[DOP-19907] - minor fixes
