+++ syncmaster/backend/dependencies/stub.py
@@ -1,5 +1,3 @@
-# SPDX-FileCopyrightText: 2023-2024 MTS PJSC
-# SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: 2023 Andrey Tikhonov (@Tishka17)
 # SPDX-License-Identifier: Apache-2.0
