+++ syncmaster/settings/__init__.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 from enum import StrEnum
 
-from pydantic import Field, model_validator
+from pydantic import Field
 from pydantic.types import ImportString
 from pydantic_settings import BaseSettings
 
@@ -42,8 +42,7 @@ class Settings(BaseSettings):
         SYNCMASTER__SERVER__DEBUG=True
     """
 
-    PROJECT_NAME: str = "SyncMaster"
-
+    # TODO: move settings to corresponding classes
     SECRET_KEY: str = "secret"
     SECURITY_ALGORITHM: str = "HS256"
     CRYPTO_KEY: str = "UBgPTioFrtH2unlC4XFDiGf5sYfzbdSf_VgiUSaQc94="
@@ -63,48 +62,3 @@ class Settings(BaseSettings):
     class Config:
         env_prefix = "SYNCMASTER__"
         env_nested_delimiter = "__"
-
-
-class TestSettings(BaseSettings):
-    TEST_POSTGRES_HOST_FOR_CONFTEST: str
-    TEST_POSTGRES_PORT_FOR_CONFTEST: int
-    TEST_POSTGRES_HOST_FOR_WORKER: str
-    TEST_POSTGRES_PORT_FOR_WORKER: int
-    TEST_POSTGRES_DB: str
-    TEST_POSTGRES_USER: str
-    TEST_POSTGRES_PASSWORD: str
-
-    TEST_ORACLE_HOST_FOR_CONFTEST: str
-    TEST_ORACLE_PORT_FOR_CONFTEST: int
-    TEST_ORACLE_HOST_FOR_WORKER: str
-    TEST_ORACLE_PORT_FOR_WORKER: int
-    TEST_ORACLE_USER: str
-    TEST_ORACLE_PASSWORD: str
-    TEST_ORACLE_SID: str | None = None
-    TEST_ORACLE_SERVICE_NAME: str | None = None
-
-    TEST_HIVE_CLUSTER: str
-    TEST_HIVE_USER: str
-    TEST_HIVE_PASSWORD: str
-
-    TEST_HDFS_HOST: str
-    TEST_HDFS_WEBHDFS_PORT: int
-    TEST_HDFS_IPC_PORT: int
-
-    TEST_S3_HOST_FOR_CONFTEST: str
-    TEST_S3_PORT_FOR_CONFTEST: int
-    TEST_S3_HOST_FOR_WORKER: str
-    TEST_S3_PORT_FOR_WORKER: int
-    TEST_S3_BUCKET: str
-    TEST_S3_ACCESS_KEY: str
-    TEST_S3_SECRET_KEY: str
-    TEST_S3_PROTOCOL: str = "http"
-    TEST_S3_ADDITIONAL_PARAMS: dict = {}
-
-    @model_validator(mode="before")
-    def check_sid_and_service_name(cls, values):
-        sid = values.get("TEST_ORACLE_SID")
-        service_name = values.get("TEST_ORACLE_SERVICE_NAME")
-        if (sid is None) == (service_name is None):
-            raise ValueError("Connection must have one param: sid or service name")
-        return values
